it works :)

+++ data/src/main/java/com/shahad/app/data/local/entities/CharacterEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class CharacterEntity(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val name: String,
     val imageUrl: String,
     val description: String?,
