+++ useCases/src/main/java/com/shahad/app/usecases/DeleteSeriesFromFavoriteUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteSeriesFromFavoriteUseCase @Inject constructor(
     private val seriesRepository: SeriesRepository
 ) {
-    suspend operator fun invoke(seriesId: Int)=
+    suspend operator fun invoke(seriesId: Long)=
         seriesRepository.deleteFavouriteSeries(seriesId)
 
     suspend fun clearFavouriteSeries() = seriesRepository.clearFavoriteSeries()
