+++ viewModels/src/main/java/com/shahad/app/viewmodels/HomeViewModel.kt
@@ -40,7 +40,7 @@ class HomeViewModel @Inject constructor(
         }
     }
 
-    fun deleteSeriesToFavorite(seriesId: Int){
+    fun deleteSeriesToFavorite(seriesId: Long){
         viewModelScope.launch {
             deleteSeriesFromFavoriteUseCase.invoke(seriesId)
         }
